+++ src/Extension.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Dhii\Services;
 
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 
 /**
@@ -24,6 +25,8 @@ use Psr\Container\ContainerInterface;
  * ```
  *
  * @see Factory For a similar implementation that does not accept a previous service value.
+ *
+ * @phpstan-import-type ServiceRef from Service
  */
 class Extension extends Service
 {
@@ -35,6 +38,7 @@ class Extension extends Service
     /**
      * @inheritDoc
      *
+     * @param array<ServiceRef> $dependencies A list of dependencies.
      * @param callable $definition The extension definition.
      */
     public function __construct(array $dependencies, callable $definition)
@@ -45,8 +49,13 @@ class Extension extends Service
 
     /**
      * @inheritDoc
+     *
+     * @param mixed $prev The original value, if any.
+     *
+     * @throws ContainerExceptionInterface If problem resolving from container.
      */
-    public function __invoke(ContainerInterface $c, $prev = null)
+    #[\Override]
+    public function __invoke(ContainerInterface $c, mixed $prev = null): mixed
     {
         $deps = $this->resolveDeps($c, $this->dependencies);
         array_unshift($deps, $prev);