+++ src/Factories/Alias.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dhii\Services\Factories;
 
 use Dhii\Services\Service;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -30,8 +31,7 @@ use Psr\Container\NotFoundExceptionInterface;
  */
 class Alias extends Service
 {
-    /** @var string */
-    protected $key;
+    protected string $key;
 
     /** @var callable|null */
     protected $default;
@@ -54,7 +54,10 @@ class Alias extends Service
 
     /**
      * @inheritDoc
+     *
+     * @throws ContainerExceptionInterface If problem retrieving aliased key from container.
      */
+    #[\Override]
     public function __invoke(ContainerInterface $c)
     {
         if (!$c->has($this->key) && $this->default !== null) {