+++ src/Factories/GlobalVar.php
@@ -24,8 +24,7 @@ use Psr\Container\ContainerInterface;
  */
 class GlobalVar extends Service
 {
-    /** @var string */
-    protected $name;
+    protected string $name;
 
     /**
      * Constructor.
@@ -42,6 +41,7 @@ class GlobalVar extends Service
     /**
      * @inheritDoc
      */
+    #[\Override]
     public function __invoke(ContainerInterface $c)
     {
         global ${$this->name};