+++ src/Factories/ServiceList.php
@@ -6,6 +6,7 @@ namespace Dhii\Services\Factories;
 
 use Dhii\Services\ResolveKeysCapableTrait;
 use Dhii\Services\Service;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 
 /**
@@ -54,7 +55,10 @@ class ServiceList extends Service
 
     /**
      * @inheritDoc
+     *
+     * @throws ContainerExceptionInterface If problem resolving from container.
      */
+    #[\Override]
     public function __invoke(ContainerInterface $c)
     {
         return $this->resolveDeps($c, $this->dependencies);