+++ src/Factory.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dhii\Services;
 
 use Dhii\Services\Factories\Constructor;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 
 /**
@@ -22,6 +23,8 @@ use Psr\Container\ContainerInterface;
  *
  * @see   Constructor For a similar implementation that automatically injects dependencies into constructors.
  * @see   Extension For a similar implementation that can be used with extension services.
+ *
+ * @psalm-import-type ServiceRef from Service
  */
 class Factory extends Service
 {
@@ -33,6 +36,7 @@ class Factory extends Service
     /**
      * @inheritDoc
      *
+     * @param array<ServiceRef> $dependencies A list of dependencies.
      * @param callable $definition The factory definition.
      */
     public function __construct(array $dependencies, callable $definition)
@@ -44,7 +48,10 @@ class Factory extends Service
 
     /**
      * @inheritDoc
+     *
+     * @throws ContainerExceptionInterface If problem resolving from container.
      */
+    #[\Override]
     public function __invoke(ContainerInterface $c)
     {
         $deps = $this->resolveDeps($c, $this->dependencies);